feat: add tests for delete route

+++ src/todoRouter.ts
@@ -281,19 +281,22 @@ export default class TodoRouter {
    */
   deleteTodo(req: Request, res: Response) {
     const id = Number(req.params.id);
-
-    if (!id && id !== 0) {
-      res.status(400).send({
-        message: 'The type of the provided ID is invalid. Must be a number',
+    let savedTodo = this.db.getTodos({ id })[0];
+    if (!savedTodo) {
+      return res.status(404).send({
+        message: 'No Todo was found for the provided id',
+        id,
       });
     }
 
     try {
       this.db.deleteTodo(id);
-      res.status(204).send({ message: 'Todo was deleted with success' });
+      return res.status(200).send({ message: 'Todo was deleted with success' });
     } catch (error) {
       if (error instanceof DatabaseError) {
         res.status(404).send({ message: 'Todo to delete was not found' });
+      } else {
+        res.status(500).send({ message: 'Internal server error', error });
       }
     }
   }
